requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.6.0
python-multipart>=0.0.5
pytest>=7.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=1.9.0
sqlalchemy>=1.4.0
pandas>=1.3.0
pymysql>=1.0.0
requests>=2.31.0
prometheus_client>=0.17.0
streamlit>=1.28.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
cryptography>=45.0.0 